Eliminate DiscrimNet and have constructors take instantiated reward/policy objects (#354)
* Introduce DemonstrationAlgorithm base class, start fitting things into that interface

* Make MCE IRL a DemonstrationAlgorithm; add support for discounting; general tidying

* Add back support for non-stationary density baseline

* Make Adversarial and BC work again

* Add test cases for make_data_loader and add further input validation

* Move type checking earlier (fails more often in my experience)

* train_adversarial bugfix: s/expert_data/demonstrations

* Configure darglint. Lots of doc errors caught.

* Density reward algorithm: merge DensityReward and DensityTrainer into DensityAlgorithm; fix up docstrings

* Bugfix: train reward model before policy

* Fix algorithm docstrings (darglint now passes)

* Fix quickstart guide

* Trailing commas

* make_data_loader: check batch size for each batch

* Add fixed horizon check to density baseline policy training

* MCE IRL bugfix: normalize demo_state_om

* base: check batch size for each iteration

* Self-review

* Bugfixes

* Fix train_bc

* Disable darglint from flake8, still run on pre-commit/CircleCI

* Use flake8 rather than darglint directly

* Fix most darglint errors

* Add docstrings in util and data

* Make algorithms kwargs-only

* Incorporate code review suggestions

* Minor bugfixes

* Fix test

* Remove stray trailing comma

* Increase darglint strictness

* Add some missing docstrings

* Add missing typing import

* Fix all pydocstyle errors

* Fix all darglint comments

* Fix black format

* Fix indentation breaking Sphinx build

* Fix Gym related bug

* Try Gym pin another way

* Fix test_scripts ValueError->FileNotFoundError

* Make module docstring for tests consistent

* Refactor adversarial into new package including discrim nets

* Do not try to lint deleted files

* Use default dtype rather than hardcoding float32

* Move tests

* Pass initialized discrim/reward network/policies in all algorithms

* Remove unused code

* train_adversarial: bugfix reward_net default

* Apply suggestions from code review

* Make DiscrimNetGAIL use reward_net and simplify resulting code

* Remove dead preprocessing code

* Bugfix: remove discrim_net_kwargs parameter from train_adversarial

* Remove old test for discrim_net_kwargs

* Remove DiscrimNet and relocate logic to AIRL/GAIL classes

* Remove DiscrimNet and relocate logic to AIRL/GAIL classes

* Remove unused pytest-cases package

* Delete mountain car analysis code

* Transfer: use reward_test not non-existent discrim.pt

* Add TODO

* Apply suggestions from code review

* Fix Darglint: actually run on CI, disable elsewhere with new flag

* Add missing arguments

* Fix Darglint: actually run on CI, disable elsewhere with new flag

* Bump executor size for linting

* Split type checking into separate job

* Bump to xlarge again

* Fix flaky test
