--- src/imitation/analysis/__init__.py
@@ -1,1 +1,0 @@
-"""Modules for analysing experimental results."""
