+++ src/imitation/scripts/train_adversarial.py
@@ -6,7 +6,7 @@ Can be used as a CLI script, or the `train_and_plot` function can be called dire
 import logging
 import os
 import os.path as osp
-from typing import Any, Mapping, Optional
+from typing import Any, Mapping, Optional, Type
 
 import torch as th
 from sacred.observers import FileStorageObserver
@@ -14,6 +14,7 @@ from sacred.observers import FileStorageObserver
 from imitation.algorithms.adversarial import airl, gail
 from imitation.data import rollout, types
 from imitation.policies import serialize
+from imitation.rewards import reward_nets
 from imitation.scripts.config.train_adversarial import train_adversarial_ex
 from imitation.util import logger
 from imitation.util import sacred as sacred_util
@@ -25,7 +26,8 @@ def save(trainer, save_path):
     # We implement this here and not in Trainer since we do not want to actually
     # serialize the whole Trainer (including e.g. expert demonstrations).
     os.makedirs(save_path, exist_ok=True)
-    th.save(trainer.discrim_net, os.path.join(save_path, "discrim.pt"))
+    th.save(trainer.reward_train, os.path.join(save_path, "reward_train.pt"))
+    th.save(trainer.reward_test, os.path.join(save_path, "reward_test.pt"))
     # TODO(gleave): unify this with the saving logic in data_collect?
     # (Needs #43 to be merged before attempting.)
     serialize.save_stable_model(
@@ -53,8 +55,9 @@ def train_adversarial(
     checkpoint_interval: int,
     gen_batch_size: int,
     init_rl_kwargs: Mapping,
+    reward_net_cls: Optional[Type[reward_nets.RewardNet]],
+    reward_net_kwargs: Optional[Mapping[str, Any]],
     algorithm_kwargs: Mapping[str, Mapping],
-    discrim_net_kwargs: Mapping[str, Mapping],
 ) -> Mapping[str, Mapping[str, float]]:
     """Train an adversarial-network-based imitation learning algorithm.
 
@@ -98,6 +101,8 @@ def train_adversarial(
             is only used in sanity checks.
         init_rl_kwargs: Keyword arguments for `init_rl`, the RL algorithm initialization
             utility function.
+        reward_net_cls: Class of reward network to construct.
+        reward_net_kwargs: Keyword arguments passed to reward network constructor.
         algorithm_kwargs: Keyword arguments for the `GAIL` or `AIRL` constructor
             that can apply to either constructor. Unlike a regular kwargs argument, this
             argument can only have the following keys: "shared", "airl", and "gail".
@@ -108,10 +113,6 @@ def train_adversarial(
             to both the `AIRL` and `GAIL` constructors. Duplicate keyword argument keys
             between `algorithm_kwargs["shared"]` and `algorithm_kwargs["airl"]` (or
             "gail") leads to an error.
-        discrim_net_kwargs: Keyword arguments for the `DiscrimNet` constructor. Unlike a
-            regular kwargs argument, this argument can only have the following keys:
-            "shared", "airl", "gail". These keys have the same meaning as they do in
-            `algorithm_kwargs`.
 
     Returns:
         A dictionary with two keys. "imit_stats" gives the return value of
@@ -122,7 +123,7 @@ def train_adversarial(
 
     Raises:
         ValueError: `gen_batch_size` not divisible by `num_vec`.
-        ValueError: `discrim_net_kwargs` or `algorithm_kwargs` included unsupported key
+        ValueError: `algorithm_kwargs` included unsupported key
             (not one of "shared", "gail" or "airl").
         ValueError: Number of expert trajectories is less than `n_expert_demos`.
         FileNotFoundError: `rollout_path` does not exist.
@@ -133,14 +134,9 @@ def train_adversarial(
         )
 
     allowed_keys = {"shared", "gail", "airl"}
-    if not discrim_net_kwargs.keys() <= allowed_keys:
-        raise ValueError(
-            f"Invalid discrim_net_kwargs.keys()={discrim_net_kwargs.keys()}. "
-            f"Allowed keys: {allowed_keys}",
-        )
     if not algorithm_kwargs.keys() <= allowed_keys:
         raise ValueError(
-            f"Invalid discrim_net_kwargs.keys()={algorithm_kwargs.keys()}. "
+            f"Invalid algorithm_kwargs.keys()={algorithm_kwargs.keys()}. "
             f"Allowed keys: {allowed_keys}",
         )
 
@@ -178,16 +174,21 @@ def train_adversarial(
         **init_rl_kwargs,
     )
 
-    discrim_kwargs_shared = discrim_net_kwargs.get("shared", {})
-    discrim_kwargs_algo = discrim_net_kwargs.get(algorithm, {})
-    final_discrim_kwargs = dict(**discrim_kwargs_shared, **discrim_kwargs_algo)
-
     algorithm_kwargs_shared = algorithm_kwargs.get("shared", {})
     algorithm_kwargs_algo = algorithm_kwargs.get(algorithm, {})
     final_algorithm_kwargs = dict(
         **algorithm_kwargs_shared,
         **algorithm_kwargs_algo,
     )
+
+    reward_net = None
+    if reward_net_cls is not None:
+        reward_net_kwargs = reward_net_kwargs or {}
+        reward_net = reward_net_cls(
+            venv.observation_space,
+            venv.action_space,
+            **reward_net_kwargs,
+        )
 
     if algorithm.lower() == "gail":
         algo_cls = gail.GAIL
@@ -201,16 +202,17 @@ def train_adversarial(
         demonstrations=expert_transitions,
         gen_algo=gen_algo,
         log_dir=log_dir,
-        discrim_kwargs=final_discrim_kwargs,
+        reward_net=reward_net,
         custom_logger=custom_logger,
         **final_algorithm_kwargs,
     )
 
-    logging.info(f"Discriminator network summary:\n {trainer.discrim_net}")
+    logging.info(f"Reward network summary:\n {reward_net}")
     logging.info(f"RL algorithm: {type(trainer.gen_algo)}")
     logging.info(
         f"Imitation (generator) policy network summary:\n" f"{trainer.gen_algo.policy}",
     )
+    logging.info(f"Adversarial algorithm: {algorithm}")
 
     def callback(round_num):
         if checkpoint_interval > 0 and round_num % checkpoint_interval == 0:
